Move SavedMusic to a separated feature module
