+++ core/common/src/main/res/drawable/ic_empty_playlist.xml
