+++ feature/saved-music/src/main/res/anim/bottom_lifting_animation.xml
