+++ feature/saved-music/src/main/res/layout/fragment_saved_music.xml
@@ -3,10 +3,10 @@
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto">
 
-    <com.yurii.youtubemusic.ui.NestedScrollCoordinatorLayout
+    <com.youtubemusic.core.common.ui.NestedScrollCoordinatorLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:background="@color/white"
+        android:background="@android:color/white"
         android:clipToPadding="false">
 
         <com.google.android.material.appbar.AppBarLayout
@@ -33,5 +33,5 @@
             android:layout_height="match_parent"
             app:layout_behavior="com.google.android.material.appbar.AppBarLayout$ScrollingViewBehavior" />
 
-    </com.yurii.youtubemusic.ui.NestedScrollCoordinatorLayout>
+    </com.youtubemusic.core.common.ui.NestedScrollCoordinatorLayout>
 </layout>
