+++ feature/saved-music/src/main/res/layout/item_music.xml
@@ -21,7 +21,7 @@
         android:clickable="true"
         android:focusable="true"
         android:foreground="?android:attr/selectableItemBackground"
-        app:cardBackgroundColor="@color/white">
+        app:cardBackgroundColor="@android:color/white">
 
         <RelativeLayout
             android:layout_width="match_parent"
@@ -39,7 +39,7 @@
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
                     android:adjustViewBounds="true"
-                    app:imageSrc="@{musicItem.thumbnail}"
+                    imageSrc="@{musicItem.thumbnail}"
                     app:shapeAppearanceOverlay="@style/roundedImageView"
                     tools:ignore="ContentDescription"
                     tools:src="@tools:sample/backgrounds/scenic" />
@@ -61,7 +61,7 @@
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:layout_centerVertical="true"
-                android:layout_marginStart="@dimen/mediumMargin"
+                android:layout_marginStart="8dp"
                 android:layout_toStartOf="@id/duration"
                 android:layout_toEndOf="@id/thumbnail_layout"
                 android:orientation="vertical">
@@ -123,7 +123,7 @@
                 android:layout_height="wrap_content"
                 android:layout_alignParentEnd="true"
                 android:layout_centerVertical="true"
-                android:layout_marginEnd="@dimen/mediumMargin"
+                android:layout_marginEnd="8dp"
                 android:background="@android:color/transparent"
                 android:contentDescription="@string/content_description_more_options"
                 android:src="@drawable/ic_baseline_more_vert_24" />
