+++ feature/saved-music/src/main/res/layout/sheet_share_content.xml
@@ -11,7 +11,7 @@
         android:layout_height="100dp"
         android:layout_centerHorizontal="true"
         android:layout_marginTop="16dp"
-        app:civ_border_color="@color/colorAccent"
+        app:civ_border_color="?attr/colorAccent"
         app:civ_border_width="4dp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
