+++ settings.gradle
@@ -26,3 +26,4 @@ include ':feature:equalizer'
 include ':feature:playlist-editor'
 include ':feature:download-manager'
 include ':feature:player'
+include ':feature:saved-music'
